+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Remove unexpected `CARD_REMOVED` event processing in **Wait For Card Insertion** state.
 
 ## [2.3.5] - 2024-06-24
 ### Fixed